+++ src/script/lang.rs
@@ -354,9 +354,8 @@ impl Script {
     ///
     /// * `block_number`  - The block time to push
     pub fn new_for_coinbase(block_number: u64) -> Self {
-        Self {
-            stack: vec![StackEntry::Num(block_number as usize)],
-        }
+        let stack = vec![StackEntry::Num(block_number as usize)];
+        Self { stack }
     }
 
     /// Constructs a new script for an asset creation
@@ -373,19 +372,16 @@ impl Script {
         signature: Signature,
         pub_key: PublicKey,
     ) -> Self {
-        let mut new_script = Script::new();
-
-        new_script.stack.push(StackEntry::Op(OpCodes::OP_CREATE));
-        new_script
-            .stack
-            .push(StackEntry::Num(block_number as usize));
-        new_script.stack.push(StackEntry::Op(OpCodes::OP_DROP));
-        new_script.stack.push(StackEntry::Bytes(asset_hash));
-        new_script.stack.push(StackEntry::Signature(signature));
-        new_script.stack.push(StackEntry::PubKey(pub_key));
-        new_script.stack.push(StackEntry::Op(OpCodes::OP_CHECKSIG));
-
-        new_script
+        let stack = vec![
+            StackEntry::Op(OpCodes::OP_CREATE),
+            StackEntry::Num(block_number as usize),
+            StackEntry::Op(OpCodes::OP_DROP),
+            StackEntry::Bytes(asset_hash),
+            StackEntry::Signature(signature),
+            StackEntry::PubKey(pub_key),
+            StackEntry::Op(OpCodes::OP_CHECKSIG),
+        ];
+        Self { stack }
     }
 
     /// Constructs a pay to public key hash script
@@ -401,28 +397,22 @@ impl Script {
         pub_key: PublicKey,
         address_version: Option<u64>,
     ) -> Self {
-        let mut new_script = Script::new();
-        let pub_key_stack_entry = StackEntry::PubKey(pub_key);
-        let new_key = construct_address_for(&pub_key, address_version);
-
         let op_hash_256 = match address_version {
             Some(NETWORK_VERSION_V0) => OpCodes::OP_HASH256_V0,
             Some(NETWORK_VERSION_TEMP) => OpCodes::OP_HASH256_TEMP,
             _ => OpCodes::OP_HASH256,
         };
-
-        new_script.stack.push(StackEntry::Bytes(check_data));
-        new_script.stack.push(StackEntry::Signature(signature));
-        new_script.stack.push(pub_key_stack_entry);
-        new_script.stack.push(StackEntry::Op(OpCodes::OP_DUP));
-        new_script.stack.push(StackEntry::Op(op_hash_256));
-        new_script.stack.push(StackEntry::PubKeyHash(new_key));
-        new_script
-            .stack
-            .push(StackEntry::Op(OpCodes::OP_EQUALVERIFY));
-        new_script.stack.push(StackEntry::Op(OpCodes::OP_CHECKSIG));
-
-        new_script
+        let stack = vec![
+            StackEntry::Bytes(check_data),
+            StackEntry::Signature(signature),
+            StackEntry::PubKey(pub_key),
+            StackEntry::Op(OpCodes::OP_DUP),
+            StackEntry::Op(op_hash_256),
+            StackEntry::PubKeyHash(construct_address_for(&pub_key, address_version)),
+            StackEntry::Op(OpCodes::OP_EQUALVERIFY),
+            StackEntry::Op(OpCodes::OP_CHECKSIG),
+        ];
+        Self { stack }
     }
 
     /// Constructs one part of a multiparty transaction script
@@ -433,14 +423,13 @@ impl Script {
     /// * `pub_key`     - Public key of this party
     /// * `signature`   - Signature of this party
     pub fn member_multisig(check_data: String, pub_key: PublicKey, signature: Signature) -> Self {
-        let mut new_script = Script::new();
-
-        new_script.stack.push(StackEntry::Bytes(check_data));
-        new_script.stack.push(StackEntry::Signature(signature));
-        new_script.stack.push(StackEntry::PubKey(pub_key));
-        new_script.stack.push(StackEntry::Op(OpCodes::OP_CHECKSIG));
-
-        new_script
+        let stack = vec![
+            StackEntry::Bytes(check_data),
+            StackEntry::Signature(signature),
+            StackEntry::PubKey(pub_key),
+            StackEntry::Op(OpCodes::OP_CHECKSIG),
+        ];
+        Self { stack }
     }
 
     /// Constructs a multisig locking script
@@ -452,25 +441,11 @@ impl Script {
     /// * `check_data`  - Data to have checked against signatures
     /// * `pub_keys`    - The constituent public keys
     pub fn multisig_lock(m: usize, n: usize, check_data: String, pub_keys: Vec<PublicKey>) -> Self {
-        let mut new_script = Script::new();
-
-        if n > pub_keys.len() || m > pub_keys.len() {
-            error!("The number of keys required for multisig is greater than the number of keys provided");
-        } else if m > n {
-            error!("Multisig requiring more keys to lock than the total number of keys");
-        } else {
-            let mut new_stack = Vec::with_capacity(3 + pub_keys.len());
-
-            new_stack.push(StackEntry::Bytes(check_data));
-            new_stack.push(StackEntry::Num(m));
-            new_stack.append(&mut pub_keys.iter().map(|e| StackEntry::PubKey(*e)).collect());
-            new_stack.push(StackEntry::Num(n));
-            new_stack.push(StackEntry::Op(OpCodes::OP_CHECKMULTISIG));
-
-            new_script.stack = new_stack;
-        }
-
-        new_script
+        let mut stack = vec![StackEntry::Bytes(check_data), StackEntry::Num(m)];
+        stack.append(&mut pub_keys.iter().map(|e| StackEntry::PubKey(*e)).collect());
+        stack.push(StackEntry::Num(n));
+        stack.push(StackEntry::Op(OpCodes::OP_CHECKMULTISIG));
+        Self { stack }
     }
 
     /// Constructs a multisig unlocking script
@@ -480,16 +455,14 @@ impl Script {
     /// * `check_data`  - Data to have signed
     /// * `signatures`  - Signatures to unlock with
     pub fn multisig_unlock(check_data: String, signatures: Vec<Signature>) -> Self {
-        let mut new_script = Script::new();
-        new_script.stack = vec![StackEntry::Bytes(check_data)];
-        new_script.stack.append(
+        let mut stack = vec![StackEntry::Bytes(check_data)];
+        stack.append(
             &mut signatures
                 .iter()
                 .map(|e| StackEntry::Signature(*e))
                 .collect(),
         );
-
-        new_script
+        Self { stack }
     }
 
     /// Constructs a multisig validation script
@@ -507,36 +480,18 @@ impl Script {
         signatures: Vec<Signature>,
         pub_keys: Vec<PublicKey>,
     ) -> Self {
-        let mut new_script = Script::new();
-
-        if n > pub_keys.len() || m > pub_keys.len() {
-            error!("The number of keys required for multisig is greater than the number of keys provided");
-        } else if m > n {
-            error!("Multisig requiring more keys to lock than the total number of keys");
-        } else {
-            new_script.stack = vec![StackEntry::Bytes(check_data)];
-
-            // Handle signatures
-            new_script.stack.append(
-                &mut signatures
-                    .iter()
-                    .map(|e| StackEntry::Signature(*e))
-                    .collect(),
-            );
-
-            new_script.stack.push(StackEntry::Num(m));
-
-            // Handle pub keys
-            new_script
-                .stack
-                .append(&mut pub_keys.iter().map(|e| StackEntry::PubKey(*e)).collect());
-            new_script.stack.push(StackEntry::Num(n));
-            new_script
-                .stack
-                .push(StackEntry::Op(OpCodes::OP_CHECKMULTISIG));
-        }
-
-        new_script
+        let mut stack = vec![StackEntry::Bytes(check_data)];
+        stack.append(
+            &mut signatures
+                .iter()
+                .map(|e| StackEntry::Signature(*e))
+                .collect(),
+        );
+        stack.push(StackEntry::Num(m));
+        stack.append(&mut pub_keys.iter().map(|e| StackEntry::PubKey(*e)).collect());
+        stack.push(StackEntry::Num(n));
+        stack.push(StackEntry::Op(OpCodes::OP_CHECKMULTISIG));
+        Self { stack }
     }
 }
 
@@ -544,100 +499,5 @@ impl From<Vec<StackEntry>> for Script {
     /// Creates a new script with a pre-filled stack
     fn from(script: Vec<StackEntry>) -> Self {
         Script { stack: script }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_is_valid_script() {
-        // empty script
-        let v = vec![];
-        let script = Script::from(v);
-        assert!(script.is_valid());
-        // script length <= 10000 bytes
-        let v = vec![StackEntry::Bytes("a".repeat(500)); 20];
-        let script = Script::from(v);
-        assert!(script.is_valid());
-        // script length > 10000 bytes
-        let v = vec![StackEntry::Bytes("a".repeat(500)); 21];
-        let script = Script::from(v);
-        assert!(!script.is_valid());
-        // # opcodes <= 201
-        let v = vec![StackEntry::Op(OpCodes::OP_1); MAX_OPS_PER_SCRIPT as usize];
-        let script = Script::from(v);
-        assert!(script.is_valid());
-        // # opcodes > 201
-        let v = vec![StackEntry::Op(OpCodes::OP_1); (MAX_OPS_PER_SCRIPT+1) as usize];
-        let script = Script::from(v);
-        assert!(!script.is_valid());
-    }
-
-    #[test]
-    fn test_is_valid_stack() {
-        // empty stack
-        let v = vec![];
-        let stack = Stack::from(v);
-        assert!(stack.is_valid());
-        // # items on interpreter stack <= 1000
-        let v = vec![StackEntry::Num(1); MAX_STACK_SIZE as usize];
-        let stack = Stack::from(v);
-        assert!(stack.is_valid());
-        // # items on interpreter stack > 1000
-        let v = vec![StackEntry::Num(1); (MAX_STACK_SIZE+1) as usize];
-        let stack = Stack::from(v);
-        assert!(!stack.is_valid());
-    }
-
-    #[test]
-    fn test_interpret_script() {
-        // empty script
-        let v = vec![];
-        let script = Script::from(v);
-        assert!(script.interpret());
-        // OP_0
-        let v = vec![StackEntry::Op(OpCodes::OP_0)];
-        let script = Script::from(v);
-        assert!(!script.interpret());
-        // OP_1
-        let v = vec![StackEntry::Op(OpCodes::OP_1)];
-        let script = Script::from(v);
-        assert!(script.interpret());
-        // OP_1 OP_2 OP_ADD OP_3 OP_EQUAL
-        let v = vec![
-            StackEntry::Op(OpCodes::OP_1),
-            StackEntry::Op(OpCodes::OP_2),
-            StackEntry::Op(OpCodes::OP_ADD),
-            StackEntry::Op(OpCodes::OP_3),
-            StackEntry::Op(OpCodes::OP_EQUAL),
-        ];
-        let script = Script::from(v);
-        assert!(script.interpret());
-        // script length <= 10000 bytes
-        let v = vec![StackEntry::Bytes("a".repeat(500)); 20];
-        let script = Script::from(v);
-        assert!(script.interpret());
-        // script length > 10000 bytes
-        let v = vec![StackEntry::Bytes("a".repeat(500)); 21];
-        let script = Script::from(v);
-        assert!(!script.interpret());
-        // # opcodes <= 201
-        let v = vec![StackEntry::Op(OpCodes::OP_1); MAX_OPS_PER_SCRIPT as usize];
-        let script = Script::from(v);
-        assert!(script.interpret());
-        // # opcodes > 201
-        let v = vec![StackEntry::Op(OpCodes::OP_1); (MAX_OPS_PER_SCRIPT+1) as usize];
-        let script = Script::from(v);
-        assert!(!script.interpret());
-        // # items on interpreter stack <= 1000
-        let v = vec![StackEntry::Num(1); MAX_STACK_SIZE as usize];
-        let script = Script::from(v);
-        assert!(script.interpret());
-        // # items on interpreter stack > 1000
-        let v = vec![StackEntry::Num(1); (MAX_STACK_SIZE+1) as usize];
-        let script = Script::from(v);
-        assert!(!script.interpret());
     }
 }
